pyproject.toml
[project]
name = "soroka"
version = "0.1.0"
description = "Telegram-bot knowledge base for forwarded content"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[ext]==22.*",
    "httpx==0.27.*",
    "pydantic==2.*",
    "trafilatura==1.12.*",
    "pypdf==4.*",
    "python-docx==1.*",
    "openpyxl==3.*",
    "pytesseract==0.3.*",
    "Pillow==10.*",
    "sqlite-vec==0.1.*",
    "mcp==1.*",
    "python-dotenv==1.*",
]

[project.optional-dependencies]
dev = [
    "pytest==8.*",
    "pytest-asyncio==0.23.*",
    "pytest-httpx==0.30.*",
    "freezegun==1.*",
    "reportlab==4.*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
